Moved utility functions from the index.js to utils.js
